collector serverless

+++ packages/watchmen-storage-postgresql/tenant-scripts/18.0.0/00008-change_data_json_history.ddl.sql
@@ -6,7 +6,7 @@ CREATE TABLE change_data_json_history
     object_id           VARCHAR(50)      NOT NULL,
     table_name          VARCHAR(50)      NOT NULL,
     data_id             JSON             NOT NULL,
-    sequence            INTEGER,
+    sequence            BIGINT,
     content             JSON,
     depend_on           JSON,
     is_posted           SMALLINT         NOT NULL,
